refactor(character): Stage C.3 — split core/character_profile.py 24KB → package, all files ≤11KB
Splits the 24 KB monolithic core/character_profile.py into a 3-mixin
package along the traits-registry / state-and-ripple / summary-and-
prompt seam. Resolves PLATFORM_READINESS Gate v0.3 dimension A for
this file (3rd of five violations after C.1 wiki_generator and
C.2 cascade).

## What changed

- core/character_profile.py (24 KB, 444 lines, 1 class with 10
  methods + 2 module-level helpers) → core/character_profile/
  package preserving the external surface byte-identical.
- No method signatures, attributes, return shapes, or behaviour
  changed. Split is purely textual.

## File sizes (post-split)

  __init__.py    2.1 KB   facade, class composition, get_profile singleton
  _traits.py     9.3 KB   16-trait registry + 28-edge correlations + _CORR_INDEX
  _profile.py    5.7 KB   _ProfileCoreMixin (state, get/set, ripple, _load/_save)
  _summary.py   10.6 KB   _SummaryMixin (build_summary + get_prompt_modifiers)

All four files comfortably under the 20 KB gate.

## Mixin composition

  CharacterProfile(_SummaryMixin, _ProfileCoreMixin)

MRO: Summary → ProfileCore → object. No method-name collisions; order
is documentation-driven (read/projection layer above state layer).

## External surface preserved

All current import paths work byte-identical:

- from core.character_profile import CharacterProfile        (5+ sites)
- from core.character_profile import get_profile             (server_llmwiki.py:2249, 2268)
- from core.character_profile import TRAITS                  (tests/test_character_profile_i18n.py)
- from core.character_profile import TRAITS, CharacterProfile (test_character_summary.py)
- from core.character_profile import (
    TRAITS, CORRELATIONS, CharacterProfile,
    _CORR_INDEX, _OPPONENTS,
  )                                                          (test_character_traits_correlations.py)
- from core.character_profile import (
    CORRELATIONS, CharacterProfile, _RIPPLE_DAMPING,
  )                                                          (test_character_w3b.py)

The package __init__.py re-exports every name including private
constants that tests pin.

## Verification

- character_profile-specific tests pass — 96 passed + 32 subtests
  across i18n / summary / traits_correlations / w3b / remove_text
  _persona.
- Full pytest suite (matching CI ignore list): 1590 passed + 824
  subtests passed. Zero regression.
- scripts/dogfood_check.py — 4/4 plugin invariants still PASS.
- ruff check core/character_profile/ — clean.

## Stage C progress

- C.1 (wiki_generator.py 51 KB) — ✅ PR #427
- C.2 (cascade.py 24 KB) — ✅ PR #428
- C.3 (character_profile.py 24 KB) — ✅ this PR
- C.4 (security_layer.py 23 KB) — pending (after Stage D.1 JSONL writers)
- C.5 (graph_editor.py 20 KB boundary) — pending; reassess after C.4

3 of 5 module-size violations cleared on origin/main.

## Out of scope

- Trait registry content changes — pure refactor.
- DB schema or persistence behaviour — _load / _save identical.
